Update to latest serenity release and begin work on web interface

+++ src/audio/discord_audio_packet.rs
@@ -1,17 +1,17 @@
-pub static PACKET_INTERVAL: u64 = 20;
-pub static PACKET_SIZE: usize = 1920;
+pub static PACKET_INTERVAL: u8 = 20;
+pub static PACKET_SIZE: u16 = 1920;
 
 #[derive(Clone)]
 pub struct DiscordAudioPacket {
     pub ssrc: u32,
     pub sequence: u16,
-    pub timestamp: u64,
+    pub timestamp: u32,
     pub stereo: bool,
     pub data: Vec<i16>,
 }
 
 impl DiscordAudioPacket {
-    pub fn new(ssrc: u32, sequence: u16, timestamp: u64, stereo: bool, data: Vec<i16>) -> Self {
+    pub fn new(ssrc: u32, sequence: u16, timestamp: u32, stereo: bool, data: Vec<i16>) -> Self {
         Self {
             ssrc,
             sequence,
